Move EditorActions connect
