+++ ui/src/components/index.js
@@ -9,3 +9,4 @@ export { Plate } from './editor/Plate';
 export { NoPlatesMessage } from './editor/Plate';
 export { ComponentList } from './editor/ComponentList';
 export { Tools } from './editor/Tools';
+export { EditorActions } from './editor/EditorActions';
