+++ ui/src/implementations/kapture-backend/components/index.js
@@ -9,3 +9,4 @@ export { PlateTabBar } from './PlateTabBarConnected';
 export { Plate } from './PlateConnected';
 export { ComponentList } from './ComponentListConnected';
 export { Tools } from './ToolsConnected';
+export { EditorActions } from './EditorActionsConnected';
